fetch integrations after clearing out previous accounts integrations
